rename ray_register to ray_remote
