+++ cpp/include/ray/api.h
@@ -6,7 +6,7 @@
 #include <ray/api/actor_task_caller.h>
 #include <ray/api/exec_funcs.h>
 #include <ray/api/object_ref.h>
-#include <ray/api/ray_register.h>
+#include <ray/api/ray_remote.h>
 #include <ray/api/ray_runtime.h>
 #include <ray/api/ray_runtime_holder.h>
 #include <ray/api/task_caller.h>
